grain pelette smart internz project final/requirements.txt
flask
tensorflow-cpu==2.15.0
numpy
Pillow
